admin crud almost done

+++ src/controllers/users.ts
@@ -11,7 +11,6 @@ class Controller {
   list = async (req: Request, res: Response) => {
     try {
       const data: User = await validators.superAdmin.validateAsync(req.user);
-      console.log(data);
       if (!((await data.isSuperAdmin) == true)) {
         res.send(401).send();
       }
@@ -39,23 +38,46 @@ class Controller {
 
   update = async (req: Request, res: Response) => {
     try {
-      const data: User = await validators.updateUser.validateAsync(req.body);
+      const data: User = await validators.superAdmin.validateAsync(req.user);
+      if (!((await data.isSuperAdmin) == true)) {
+        return res.send(401).send("unauthorized");
+      }
+      const userdata: User = await validators.updateUser.validateAsync(
+        req.body
+      );
+      //do i put a findUnique user to get a 404?
+      const user = await prisma.user.update({
+        where: {
+          id: req.params.id,
+        },
+        data: { ...userdata },
+      });
+      if (!user) {
+        return res.status(404).send();
+      }
+      res.status(200).send(user);
     } catch (e) {
-      res.status(400);
+      res.status(400).send(e);
     }
   };
 
   destroy = async (req: Request, res: Response) => {
     try {
-      const id = req.params.id;
-      await prisma.user.delete({
+      const data: User = await validators.superAdmin.validateAsync(req.user);
+      if (!((await data.isSuperAdmin) == true)) {
+        return res.send(401).send("unauthorized");
+      }
+      const user = await prisma.user.delete({
         where: {
-          id: id,
+          id: req.params.id,
         },
       });
+      if (!user) {
+        return res.status(404).send();
+      }
       res.status(200).send("user deleted");
     } catch (e) {
-      res.status(500).send();
+      res.status(400).send();
     }
   };
 }
